backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
networkx==3.4.2
pandas==2.2.3
numpy==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
pydantic-settings==2.7.1
python-multipart==0.0.20
structlog==24.4.0
psycopg2-binary==2.9.10
